+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # 0.2.0
 - Created the `Outcome` class under roulette module
 - Create the `Bin` class as an extended collection data structure of the frozenset.
+- Create the `Wheel` class under roulette module
+
 
 # 0.1.0
 - Unit test using `unittest` class